% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables_managment.R
\name{pgDropTables}
\alias{pgDropTables}
\title{Drop tables from a PostgreSQL database}
\usage{
pgDropTables(conn, pattern, verbose = TRUE)
}
\arguments{
\item{conn}{a PostgreSQLConnection object.}

\item{pattern}{character; string pattern matching tables names .}

\item{verbose}{boolean; whether or not drop message must be printed.}
}
\description{
Drop tables from a PostgreSQL database matching a pattern with CASCADE.
}
\examples{
library(RPostgreSQL)
data(mtcars)
conn = dbConnect(drv=dbDriver("PostgreSQL"), host="localhost", port=5432,
                dbname="foodflows", user="milo", password="postgres")
dbWriteTable(conn, "mtcars", mtcars, overwrite=TRUE)
dbWriteTable(conn, "mtcars2", mtcars, overwrite=TRUE)
pgDropTables(conn, "cars") # Will drop the 2
dbDisconnect(conn)

}
\references{
https://stackoverflow.com/questions/4202135/how-to-drop-multiple-tables-in-postgresql-using-a-wildcard
}
