% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{pgisGetBbox}
\alias{pgisGetBbox}
\title{Get bbox of a table}
\usage{
pgisGetBbox(conn, tbl, geom.col = "geometry")
}
\arguments{
\item{conn}{a PostgreSQLConnection object.}

\item{tbl}{character; table name.}

\item{geom.col}{character; name of the geometry column (default: 'geometry).}
}
\value{
a bbox class object.
}
\description{
Get the extent of the geometry of a PostGIS table, in the same EPSG
than table.
}
\examples{
library(spData)
library(sf)
library(RPostgreSQL)
data(seine)
conn = dbConnect(drv=dbDriver("PostgreSQL"), host="localhost", port=5432,
                dbname="foodflows", user="milo", password="postgres")
dbWriteTable(conn, name="seine", seine, overwrite=TRUE)
bb = pgisGetBbox(conn, "seine")
(class(bb)=="bbox")
dbDisconnect(conn)

}
