% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{pgisGetEPSG}
\alias{pgisGetEPSG}
\title{Get EPSG of a table}
\usage{
pgisGetEPSG(conn, tbl, geom.col = "geometry", disconnect = TRUE)
}
\arguments{
\item{conn}{a PostgreSQLConnection object.}

\item{tbl}{character; table name.}

\item{geom.col}{character; name of the geometry column (default: 'geometry).}

\item{disconnect}{boolean; shall the connection be closed at the end ? (default: TRUE)}
}
\value{
a numeric EPSG code of 4 numbers.
}
\description{
Get EPSG of a table
}
\examples{
library(spData)
library(sf)
library(RPostgreSQL)
data(seine)
conn = dbConnect(drv=dbDriver("PostgreSQL"), host="localhost", port=5432,
                dbname="foodflows", user="postgres", password="postgres")
dbWriteTable(conn, name="seine", seine, overwrite=TRUE)
(pgisGetEPSG(conn, "seine")==2154)

}
