% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasters.R
\name{pgisGetRaster}
\alias{pgisGetRaster}
\title{Get raster from PostGIS}
\usage{
pgisGetRaster(conn, tbl, flip = FALSE, password = "postgres")
}
\arguments{
\item{conn}{a PostgreSQLConnection object.}

\item{tbl}{character; name of the raster table.}

\item{flip}{boolean; flip raster Y axis if necessary (default: FALSE).}

\item{password}{character; password for database (default: 'postgres').}
}
\value{
a RasterLayer object.
}
\description{
Get raster table from a PostGIS database. The raster can be verticaly flipped
if necessary, if a Y axis bug in rdgal occurs.
}
\examples{
library(RPostgreSQL)
library(spData)
library(raster)
data(elev)
pgisWriteRaster(elev, dbname="foodflows", crs=2154)
conn = dbConnect(drv=dbDriver("PostgreSQL"), host="localhost", port=5432,
                dbname="foodflows", user="milo", password="postgres")
rast = pgisGetRaster(conn, tbl="elev")
raster::plot(rast)

}
