% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{pgisMakeValid}
\alias{pgisMakeValid}
\title{Correct PostGIS table geometry errors}
\usage{
pgisMakeValid(
  conn,
  tbl,
  geom.col = "geometry",
  verbose = TRUE,
  disconnect = TRUE
)
}
\arguments{
\item{conn}{a PostgreSQLConnection object.}

\item{tbl}{character; table name.}

\item{geom.col}{character; name of the geometry column (default: 'geometry).}

\item{verbose}{boolean; wheter or not warning when correction are applied (default: TRUE)}

\item{disconnect}{boolean; shall the connection be closed at the end ? (default: TRUE)}
}
\description{
The geometry of a table can be invalid (most of the time because of
 intersecting segments), ans causes bugs in geoprocessing. This function
 correct these errors if necessary.
}
