% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasters.R
\name{pgisRasterizeTable}
\alias{pgisRasterizeTable}
\title{Rasterize a PostGIS vector table}
\usage{
pgisRasterizeTable(
  conn,
  geom.tbl,
  rast.tbl,
  xref = NULL,
  yref = NULL,
  scale = 1000,
  disconnect = TRUE
)
}
\arguments{
\item{conn}{a PostgreSQLConnection object.}

\item{geom.tbl}{character; name of the input vector table}

\item{rast.tbl}{character; name of the output raster table}

\item{xref}{numeric; x coordinate of reference pixel bottomleft corner.
If NULL (default), the center of the extent is taken.}

\item{yref}{numeric; y coordinate of reference pixel bottomleft corner.
If NULL (default), the center of the extent is taken.}

\item{scale}{numeric; size of the raster pixels, in meters (default: 1000).}

\item{disconnect}{boolean; shall the connection be closed at the end ? (default: TRUE)}
}
\description{
Rasterize a PostGIS vector table.
}
\examples{
library(spData)
library(RPostgreSQL)
library(sf)
data(seine)
conn = dbConnect(drv=dbDriver("PostgreSQL"), host="localhost", port=5432,
                dbname="foodflows", user="milo", password="postgres")
sf::dbWriteTable(conn, name="seine", seine, overwrite=TRUE)
pgisRasterizeTable(conn, "seine", "seine_rast", disconnect=FALSE)
rast = pgisGetRaster(conn, "seine_rast", flip=TRUE)
raster::plot(rast)

}
\references{
https://nronnei.github.io/blog/2017/03/creating-rasters-from-scratch-in-postgis/
http://geospatialelucubrations.blogspot.com/2014/05/a-guide-to-rasterization-of-vector.html
}
