% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{pgisSimplifyGeom}
\alias{pgisSimplifyGeom}
\title{Simplify geometry of a PostGIS table}
\usage{
pgisSimplifyGeom(
  conn,
  tbl,
  tolerance,
  geom.col = "geometry",
  new.geom.col = "geometry_simpl",
  verbose = TRUE
)
}
\arguments{
\item{conn}{a PostgreSQLConnection object.}

\item{tbl}{character; the table to simplify.}

\item{tolerance}{numeric; the tolerance in meters CHECK THIS}

\item{geom.col}{character; the geometry column (default; 'geometry')}

\item{new.geom.col}{character; the name of the geometry column in which new
simplified geometry will be written. If identical to \code{geom.col}, this
one is overwritten (default: geometry_simpl).}

\item{verbose}{boolean; shall the gain of memory be printed ? (default: TRUE).}
}
\description{
Highly detailed geometries use more memory and make geoprocesses longer.
If we don't need high precision, we can simplify (=generalize) it.
}
\examples{
library(spData)
library(sf)
library(RPostgreSQL)
data(seine)
conn = dbConnect(drv=dbDriver("PostgreSQL"), host="localhost", port=5432,
               dbname="foodflows", user="postgres", password="postgres")
dbWriteTable(conn, name="seine", seine, overwrite=TRUE)
pgisSimplifyGeom(conn, "seine", tolerance=10)
pgisSimplifyGeom(conn, "seine", tolerance=100)
pgisSimplifyGeom(conn, "seine", tolerance=1000)
dbDisconnect(conn)

}
