% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgisWriteGraph.R
\name{pgisWriteGraph}
\alias{pgisWriteGraph}
\title{Title}
\usage{
pgisWriteGraph(
  conn,
  what = c("vertices", "edges"),
  coords = c("x", "y"),
  crs = 2154,
  geom.col = "geometry",
  disconnect = TRUE
)
}
\arguments{
\item{conn}{a PostgreSQLConnection object.}

\item{what}{character vector; what must be written into DB: vertices and/or
edges ? Default: the 2.}

\item{coords}{names of the numeric columns holding coordinates in vertices
dataframe (default: c('x','y')).}

\item{crs}{numeric(4); EPSG code of the Coordinate Reference System.}

\item{geom.col}{character; name of the geometry column in output PostGIS table
(default: 'geometry').}

\item{disconnect}{boolean; shall the connection be closed at the end ? (default: TRUE)}
}
\description{
Title
}
