% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasters.R
\name{pgisWriteRaster}
\alias{pgisWriteRaster}
\title{Write a raster into PostGIS}
\usage{
pgisWriteRaster(rast, tbl, dbname, crs)
}
\arguments{
\item{rast}{RasterLayer object.}

\item{tbl}{character; name of the table created.}

\item{dbname}{character; name of the database.}

\item{crs}{numeric(4); EPSG code of the Coordinate Reference System.}
}
\description{
Import a RasterLayer object into a PostGIS database.
}
\examples{
library(spData)
data(elev)
pgisWriteRaster(elev, dbname="foodflows", crs=2154)

}
\references{
https://postgis.net/docs/using_raster_dataman.html
}
