% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables_managment.R
\name{pgListIndexes}
\alias{pgListIndexes}
\title{List indexes of a database}
\usage{
pgListIndexes(conn, tables)
}
\arguments{
\item{conn}{a PostgreSQLConnection object.}

\item{tables}{character vector; tables names from wich to list indexes.
If NULL (default), all indexes all returned.}
}
\value{
a dataframe.
}
\description{
List all indexes attached to tables of a database.
}
\examples{
library(RPostgreSQL)
data(mtcars)
conn = dbConnect(drv=dbDriver("PostgreSQL"), host="localhost", port=5432,
                dbname="postgres", user="milo", password="postgres")
dbWriteTable(conn, "mtcars", mtcars, overwrite=TRUE)
dbSendQuery(conn, "CREATE INDEX IF NOT EXISTS mpg_idx ON mtcars USING btree(mpg);")
head(pgListIndexes(conn, "mtcars"))
dbDisconnect(conn)

}
