% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables_managment.R
\name{pgListIndexes}
\alias{pgListIndexes}
\title{List indexes of a database}
\usage{
pgListIndexes(conn, tables, disconnect = TRUE)
}
\arguments{
\item{conn}{a PostgreSQLConnection object.}

\item{tables}{character vector; tables names from wich to list indexes.
If NULL (default), all indexes all returned.}

\item{disconnect}{boolean; shall the connection be closed at the end ? (default: TRUE)}
}
\value{
a dataframe
}
\description{
List all indices attached to tables of a database.
}
\examples{
library(RPostgreSQL)
data(mtcars)
conn = dbConnect(drv=dbDriver("PostgreSQL"), host="localhost", port=5432,
                 dbname="postgres", user="postgres", password="postgres")
dbWriteTable(conn, "mtcars", mtcars, overwrite=TRUE)
dbSendQuery(conn, "CREATE INDEX IF NOT EXISTS mpg_idx ON mtcars USING btree(mpg);")
df = pgListIndexes(conn, "mtcars") # Same

}
